% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/theme_black.R
\name{theme_black}
\alias{theme_black}
\title{theme_black}
\usage{
theme_black(base_size = 12, base_family = "")

theme_black(base_size = 12, base_family = "")
}
\description{
theme_black

theme_black
}
\examples{
ggplot() + theme_black()
ggplot() + theme_black()
}
